+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 - Fix backup task credentials for GAE Python 3
 - Fix `djangae.contrib.security.commands_utils.extract_views_from_urlpatterns` function
 - Fix error classes not inheriting from `Exception`
-
+- Fix an issue with `djangae.storage.CloudStorage` where calling `_open()` or `delete()` wouldn't use the correct bucket
 
 ### Bug fixes: